+++ database/scripts/create_dev_schema.sql
@@ -1,45 +1,76 @@
 /* */
-/* The purpose of this SQL script is to create a schema with datasets for development.*/
+/* The purpose of this SQL script is to create schemas with data for development and testing. */
+/* Data for testing is smaller. */
 /* The schema including its data can be dumped using pg_dump. */
-/* This is usefull to setup a local development database. */
-/* It is used as basis for the development database inside a Docker container */
+/* This is usefull to setup a local development or testing database. */
+/* It is used as basis for the development database inside a Docker container. */
 /* */
-SET search_path TO public, development;
+SET search_path TO public, development, test;
 
+
+/* Development */
 CREATE SCHEMA IF NOT EXISTS development;
 
-DROP TABLE IF EXISTS development.ghs_pop;
+CREATE TABLE development.regions (
+    LIKE public.regions INCLUDING INDEXES
+);
 
+CREATE TABLE development.ghs_pop (
+    LIKE public.ghs_pop INCLUDING INDEXES
+);
 
-/* */
-/* TODO: Does not work with pg_restore due to column defaults pointing to public table. */
-/* DROP TABLE IF EXISTS development.test_regions; */
-/* CREATE TABLE development.test_regions ( */
-/*     /1* LIKE public.test_regions INCLUDING ALL *1/ */
-/*     LIKE public.test_regions */
-/* ); */
-/* INSERT INTO development.test_regions */
-/* SELECT */
-/*     * */
-/* FROM */
-/*     public.test_regions; */
-/* */
-CREATE TABLE development.ghs_pop AS
+INSERT INTO development.regions
 SELECT
+    *
+FROM
+    public.regions;
+
+INSERT INTO development.ghs_pop
+SELECT DISTINCT ON (rid)
     rid,
-    ST_UNION (ST_Clip (rast, ST_Buffer (geom, 0.01), TRUE)) AS rast
+    ST_Clip (rast, ST_Buffer (geom, 0.01), TRUE) AS rast
 FROM
     public.ghs_pop,
-    public.test_regions
+    development.regions
 WHERE
     ST_Intersects (rast, geom)
     AND ST_BandIsNoData (rast) = FALSE
-GROUP BY
-    rid,
-    geom;
-
-CREATE INDEX dev_ghs_pop_st_convexhull_idx ON development.ghs_pop USING gist (ST_ConvexHull (rast));
+    AND name NOT IN ('Bangladesh', 'Dominican Republic', 'Haiti', 'Myanmar',
+	'South Sudan');
 
 SELECT
     AddRasterConstraints ('development'::name, 'ghs_pop'::name, 'rast'::name);
 
+
+/* Testing */
+CREATE SCHEMA IF NOT EXISTS test;
+
+CREATE TABLE test.regions (
+    LIKE public.regions INCLUDING INDEXES
+);
+
+CREATE TABLE test.ghs_pop (
+    LIKE public.ghs_pop INCLUDING INDEXES
+);
+
+INSERT INTO test.regions
+SELECT
+    *
+FROM
+    public.regions
+WHERE
+    ogc_fid IN (1, 2, 3, 4, 8, 9, 14, 28, 31);
+
+INSERT INTO test.ghs_pop
+SELECT DISTINCT ON (rid)
+    rid,
+    ST_Clip (rast, ST_Buffer (geom, 0.01), TRUE) AS rast
+FROM
+    public.ghs_pop,
+    test.regions
+WHERE
+    ST_Intersects (rast, geom)
+    AND ST_BandIsNoData (rast) = FALSE;
+
+SELECT
+    AddRasterConstraints ('test'::name, 'ghs_pop'::name, 'rast'::name);