+++ database/scripts/dump_dev_schema.sh
@@ -2,10 +2,9 @@
 # This scripts is used to dump schema and tables
 # for a minimal local development setup for OQT.
 
-pg_dump --schema=development --schema=admin > schema.sql
-pg_dump \
-    --table=test_regions \
-    --table=test_regions_ogc_fid_seq \
-    --table=test_regions_geom_geom_idx \
-    --table=test_regions_pkey \
-    > test-regions.sql
+psql -f remove_dev_schema.sql
+psql -f create_dev_schema.sql
+pg_dump --schema=test --schema=admin > schema.test.sql
+pg_dump --schema=development > schema.dev.sql
+gzip schema.dev.sql
+psql -f remove_dev_schema.sql