+++ workers/tests/integrationtests/test_cli.py
@@ -6,6 +6,7 @@ https://click.palletsprojects.com/en/7.x/testing/?highlight=testing
 import os
 import unittest
 
+import geojson
 from click.testing import CliRunner
 
 from ohsome_quality_analyst.cli import cli
@@ -33,20 +34,28 @@ class TestCliIntegration(unittest.TestCase):
                 "-l",
                 "building_count",
                 "-d",
-                "test_regions",
+                "regions",
                 "-f",
                 "3",
             ],
         )
-        assert result.exit_code == 0
+        self.assertEqual(result.exit_code, 0)
 
     @oqt_vcr.use_cassette()
     def testCreateReport(self):
         result = self.runner.invoke(
             cli,
-            ["create-report", "-r", "SimpleReport", "-d", "test_regions", "-f", "3"],
+            ["create-report", "-r", "SimpleReport", "-d", "regions", "-f", "3"],
         )
-        assert result.exit_code == 0
+        self.assertEqual(result.exit_code, 0)
+
+    def testGetAvailableRegions(self):
+        result = self.runner.invoke(
+            cli,
+            ["-q", "get-available-regions"],
+        )
+        self.assertEqual(result.exit_code, 0)
+        self.assertTrue(geojson.loads(result.output).is_valid)
 
 
 if __name__ == "__main__":