+++ workers/tests/integrationtests/test_indicator_ghs_pop_comparison_buildings.py
@@ -1,10 +1,9 @@
 import asyncio
-import os
 import unittest
 
-import geojson
 from asyncpg import Record
 
+from ohsome_quality_analyst.geodatabase import client as db_client
 from ohsome_quality_analyst.indicators.ghs_pop_comparison_buildings.indicator import (
     GhsPopComparisonBuildings,
 )
@@ -14,15 +13,12 @@ from .utils import oqt_vcr
 
 class TestIndicatorGhsPopComparisonBuildings(unittest.TestCase):
     def setUp(self):
-        infile = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "fixtures",
-            "heidelberg_altstadt.geojson",
-        )
-        with open(infile, "r") as f:
-            bpolys = geojson.load(f)
+        dataset = "regions"
+        feature_id = 31
+        self.bpolys = asyncio.run(db_client.get_bpolys_from_db(dataset, feature_id))
+        self.layer_name = "building_count"
         self.indicator = GhsPopComparisonBuildings(
-            bpolys=bpolys, layer_name="building_count"
+            bpolys=self.bpolys, layer_name=self.layer_name
         )
 
     @oqt_vcr.use_cassette()