+++ workers/tests/integrationtests/test_indicator_mapping_saturation.py
@@ -38,8 +38,8 @@ class TestIndicatorMappingSaturation(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def testFloatDivisionByZeroError(self):
         layer_name = "building_count"
-        dataset = "test_regions"
-        feature_id = 30
+        dataset = "regions"
+        feature_id = 31
         bpolys = asyncio.run(db_client.get_bpolys_from_db(dataset, feature_id))
 
         indicator = MappingSaturation(layer_name=layer_name, bpolys=bpolys)
@@ -50,8 +50,8 @@ class TestIndicatorMappingSaturation(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def testCannotConvertNanError(self):
         layer_name = "building_count"
-        dataset = "test_regions"
-        feature_id = 2
+        dataset = "regions"
+        feature_id = 28
         bpolys = asyncio.run(db_client.get_bpolys_from_db(dataset, feature_id))
 
         indicator = MappingSaturation(layer_name=layer_name, bpolys=bpolys)