+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Breaking Changes
 
 - Database schema changes: Add timestamp to indicator results !125
+- Database schema changes of the regions table !120
+    - Rename attributes `infile` to `name`
+    - Change geometry type from polygon to multipolygon
 
 ### Other Changes
 
@@ -18,6 +21,15 @@
 - Tidy up repository !138 !120
 - Return a GeoJSON when computing an indicator from the CLI !140
 - Integrate [VCR.py](https://vcrpy.readthedocs.io) to cache data for tests !133
+- Changes to available regions for pre-computed results !120:
+    - Remove fid attribute from GeoJSON Feature object properties and add id attribute to GeoJSON Feature object
+    - Rename test_regions to regions
+    - Extent regions with four countries (#196)
+    - Retrieve available regions through API and CLI
+    - Correct geometry of following duplicated regions: id 2 and id 28
+    - Remove and download regions.geojson instead
+    - Website will use regions.geojson when present. Otherwise, it will use the API endpoint.
+    - Database can be setup with available regions for development or with only regions for testing
 
 ## 0.2.0